+++ Model/Hierarchy/Attribute/Handler/HierarchyId.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2022 Hawksearch (www.hawksearch.com) - All Rights Reserved
+ * Copyright (c) 2024 Hawksearch (www.hawksearch.com) - All Rights Reserved
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -14,19 +14,19 @@ declare(strict_types=1);
 
 namespace HawkSearch\EsIndexing\Model\Hierarchy\Attribute\Handler;
 
-use HawkSearch\EsIndexing\Model\Indexing\AttributeHandlerInterface;
-use Magento\Catalog\Api\Data\CategoryInterface;
-use Magento\Framework\DataObject;
+use HawkSearch\Connector\Compatibility\PublicContractDeprecation;
 
-class HierarchyId implements AttributeHandlerInterface
+PublicContractDeprecation::triggerClassDeprecationMessage(
+    HierarchyId::class,
+    '0.7.0',
+    \HawkSearch\EsIndexing\Model\Hierarchy\Field\Handler\HierarchyId::class,
+    'In favour of a new Field Handlers logic.'
+);
+
+/**
+ * @deprecated 0.7.0 In favour of a new Field Handlers logic
+ * @see \HawkSearch\EsIndexing\Model\Hierarchy\Field\Handler\HierarchyId
+ */
+class HierarchyId extends \HawkSearch\EsIndexing\Model\Hierarchy\Field\Handler\HierarchyId
 {
-
-    /**
-     * @inheritDoc
-     * @param CategoryInterface $item
-     */
-    public function handle(DataObject $item, string $attributeCode)
-    {
-        return (int)$item->getId();
-    }
 }